added Ported the applications to Windows and fixed some incompatibilities. There are still LAPACK linking problems that I'll ask Synge to take care of.
git-svn-id: svn://localhost:7744/trunk/alps@3616 3665984a-f8b0-4bb5-b30c-cdcaecbe7ced

+++ applications/qmc/worms/Wdostep.C
@@ -57,7 +57,7 @@ void WRun::dostep()
       make_worm();
   else {
     int worm_num=0;
-    for (int64_t length=0;length<=worms_per_kink*num_kinks;++worm_num) {
+    for (long long length=0;length<=worms_per_kink*num_kinks;++worm_num) {
       length+=make_worm();
     }
     worms_per_update=0.99*worms_per_update+0.01*worm_num;
